Clean up the confidence interval framework.
* Store ci_group_size as part of the forest to avoid a mismatch between training and prediction.
* Add an estimate_variance option to prediction, to replace the fragile check ci_group_size > 1.
* Move the check for consistent ci_group_size during merges into C++.
* Use size_t instead of uint for ci_group_size.

+++ core/src/prediction/OptimizedPredictionStrategy.h
@@ -66,7 +66,7 @@ public:
   virtual std::vector<double> compute_variance(
       const std::vector<double>& average_prediction_values,
       const PredictionValues& leaf_prediction_values,
-      uint ci_group_size) = 0;
+      size_t ci_group_size) = 0;
 
  /**
   * The number of types of precomputed prediction values. For regression
