+++ core/src/prediction/QuantilePredictionStrategy.h
@@ -41,7 +41,7 @@ public:
       std::vector<std::vector<size_t>> samples_by_tree,
       std::unordered_map<size_t, double> weights_by_sampleID,
       const Observations& observations,
-      uint ci_group_size);
+      size_t ci_group_size);
 
 private:
   std::vector<double> compute_quantile_cutoffs(const std::unordered_map<size_t, double>& weights_by_sample,
