+++ core/src/prediction/RegressionPredictionStrategy.h
@@ -38,7 +38,7 @@ public:
   std::vector<double> compute_variance(
       const std::vector<double>& average,
       const PredictionValues& leaf_values,
-      uint ci_group_size);
+      size_t ci_group_size);
 
   std::vector<double> compute_debiased_error(
       size_t sample,
